Move files to a specific package
