+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/jetpack/benefits/JetpackBenefitsDialog.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.jetpack
+package com.woocommerce.android.ui.jetpack.benefits
 
 import android.os.Bundle
 import android.view.View
@@ -6,14 +6,16 @@ import android.widget.Toast
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.woocommerce.android.R
-import com.woocommerce.android.analytics.AnalyticsEvent
+import com.woocommerce.android.R.layout
+import com.woocommerce.android.R.string
+import com.woocommerce.android.R.style
+import com.woocommerce.android.analytics.AnalyticsEvent.JETPACK_INSTALL_BUTTON_TAPPED
 import com.woocommerce.android.analytics.AnalyticsTracker
+import com.woocommerce.android.analytics.AnalyticsTracker.Companion
 import com.woocommerce.android.databinding.DialogJetpackBenefitsBinding
 import com.woocommerce.android.extensions.navigateSafely
 import com.woocommerce.android.ui.base.UIMessageResolver
-import com.woocommerce.android.ui.jetpack.JetpackBenefitsViewModel.StartApplicationPasswordsInstallation
-import com.woocommerce.android.ui.jetpack.JetpackBenefitsViewModel.StartJetpackCPInstallation
+import com.woocommerce.android.ui.jetpack.JetpackBenefitsDialogDirections
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowSnackbar
 import com.woocommerce.android.widgets.CustomProgressDialog
 import dagger.hilt.android.AndroidEntryPoint
@@ -21,7 +23,7 @@ import org.wordpress.android.util.DisplayUtils
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class JetpackBenefitsDialog : DialogFragment(R.layout.dialog_jetpack_benefits) {
+class JetpackBenefitsDialog : DialogFragment(layout.dialog_jetpack_benefits) {
     companion object {
         private const val TABLET_LANDSCAPE_WIDTH_RATIO = 0.35f
         private const val TABLET_LANDSCAPE_HEIGHT_RATIO = 0.8f
@@ -36,12 +38,12 @@ class JetpackBenefitsDialog : DialogFragment(R.layout.dialog_jetpack_benefits) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         // Use fullscreen style for all cases except tablet in landscape mode
-        setStyle(STYLE_NO_TITLE, if (isTabletLandscape()) R.style.Theme_Woo_Dialog else R.style.Theme_Woo)
+        setStyle(STYLE_NO_TITLE, if (isTabletLandscape()) style.Theme_Woo_Dialog else style.Theme_Woo)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         // Specify transition animations
-        dialog?.window?.attributes?.windowAnimations = R.style.Woo_Animations_Dialog
+        dialog?.window?.attributes?.windowAnimations = style.Woo_Animations_Dialog
 
         setupObservers()
 
@@ -52,7 +54,7 @@ class JetpackBenefitsDialog : DialogFragment(R.layout.dialog_jetpack_benefits) {
         binding.installJetpackButton.setOnClickListener {
             viewModel.onInstallClick()
             AnalyticsTracker.track(
-                stat = AnalyticsEvent.JETPACK_INSTALL_BUTTON_TAPPED,
+                stat = JETPACK_INSTALL_BUTTON_TAPPED,
                 properties = mapOf(AnalyticsTracker.KEY_JETPACK_INSTALLATION_SOURCE to "benefits_modal")
             )
         }
@@ -64,8 +66,8 @@ class JetpackBenefitsDialog : DialogFragment(R.layout.dialog_jetpack_benefits) {
                 if (progressDialog?.isVisible == true) return@observe
                 progressDialog?.dismiss()
                 progressDialog = CustomProgressDialog.show(
-                    getString(R.string.jetpack_benefits_fetching_status),
-                    getString(R.string.please_wait)
+                    getString(string.jetpack_benefits_fetching_status),
+                    getString(string.please_wait)
                 ).also {
                     it.isCancelable = false
                     it.show(childFragmentManager, CustomProgressDialog.TAG)
@@ -116,5 +118,5 @@ class JetpackBenefitsDialog : DialogFragment(R.layout.dialog_jetpack_benefits) {
     }
 
     private fun isTabletLandscape() = (DisplayUtils.isTablet(context) || DisplayUtils.isXLargeTablet(context)) &&
-        DisplayUtils.isLandscape(context)
+            DisplayUtils.isLandscape(context)
 }
